+++ R/check_documentation.R
@@ -88,7 +88,7 @@ check_documentation <- function(x = ".", quiet = FALSE) {
 
   namespace <- readLines(rd_files[1])
   namespace[grepl("^export.*\\(", namespace)] |>
-    gsub(pattern = "export.*\\((.*)\\)", replacement = "\\1") -> exported
+    gsub(pattern = "export.*\\(\"?(.*?)\"?\\)", replacement = "\\1") -> exported
   vapply(rd_files[-1], rd_extract_function, vector("list", 1L)) |>
     unlist() |>
     unique() |>
@@ -96,6 +96,7 @@ check_documentation <- function(x = ".", quiet = FALSE) {
   unexported <- documented[!documented %in% exported]
   datasets <- data(package = desc(x$get_path)$get_field("Package"))
   unexported <- unexported[!unexported %in% datasets$results[, "Item"]]
+  unexported <- unexported[unexported != "reexports"]
   paste(unexported, collapse = ", ") |>
     sprintf(fmt = "documented but unexported functions: %s") -> doc_warnings
   doc_warnings <- doc_warnings[length(unexported) > 0]