+++ NEWS.md
@@ -1,3 +1,14 @@
+# checklist 0.3.6
+
+* Add an `organisation` class to store organisation rules different from those
+  of the Research Institute for Nature and Forest (INBO).
+  See `vignette("organisation", package = "checklist")` for more information.
+* `add_badges()` helps to add badges to the `README`.
+* Put double quotes around the title and abstract fields of `CITATION.cff`.
+* `check_documentation()` handles assignment functions and re-exported functions
+  correctly.
+* `check_lintr()` ignores `renv` subdirectories (#118).
+
 # checklist 0.3.5
 
 * Fix release GitHub Action.
@@ -119,7 +130,7 @@
 
 # checklist 0.2.3
 
-* Add `vignette("zenodo")` on how to set up the integration with [Zenodo](https://www.zenodo.org) and
+* Add `vignette("zenodo")` on how to set up the integration with [Zenodo](https://zenodo.org) and
   [ORCID](https://orcid.org)
 * `check_environment()` makes sure that the required repository secrets are set.
   `check_package()` performs this check when it runs in a GitHub Action.